refactor: remove deprecated graph views and simplify network graph
- Remove create-graph and create-individual pages
- Remove network-pane and related components
- Simplify graph page to redirect directly to network view
- Remove save network view dialog and export menu
- Update knowledge graph sidebar to reflect new routes and remove unused code
- Clean up imports and unused state in network page

This refactor cleans up deprecated views and streamlines the network graph experience.
